give the transport-neutral queue policy its own header
MIDIQueueManager moves from midi_cc_policy.h to a new midi_queue_policy.h, included by both
midi_cc_policy.h and midi_queue_transports.h. That breaks the traits/manager include cycle exactly as
before while leaving midi_cc_policy.h about CC policy rather than also carrying classify_message,
validate_head_message_pop and the traversal result enums, and it leaves midi_queue_manager.h closer to
the end state the header split wants.

Correct the comment on next_cc, which was wrong on both halves. Stopping at a non-CC element is not
"exactly what the old DIN Invalid path did": that aborted the whole traversal for the slot, where this
reports NoMore and lets the scheduler pop from the prefix it already found. Nor does it degrade to FIFO.
Spell out the actual old-vs-new behaviour per transport, including that old DIN only reported Invalid for
a byte it could not decode and skipped decodable non-CC messages by their own length.

Delete the three MIDIQueueStorage lane-indexed wrappers left with no callers when the matching device
state forwarders went: read_at, overwrite_at and remove_span_via_head_swap. The MIDIQueueLane methods of
the same names stay - the CC lane policy and the lane tests use them.

Move the USB drain tests to their own file beside the DIN ones rather than leaving them in a file whose
header says it only covers transport-neutral, header-only pieces.
